+++ backend/app/schemas/plan.py
@@ -1,10 +1,11 @@
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class PlanCreate(BaseModel):
     name: str
-    price: float
+    price: Decimal
     interval: str
     user_quota: int
 
@@ -12,7 +13,7 @@ class PlanCreate(BaseModel):
 class PlanOut(BaseModel):
     id: str
     name: str
-    price: float
+    price: Decimal
     interval: str
     user_quota: int
     created_at: datetime
